Landing page
We created a React component called SearchField to integrate a GeoSearchControl into a Leaflet map.
Initially, we attempted to use a custom marker icon from an imported image, but faced issues with rendering.
We then simplified by using a green dot as a marker icon using a L.divIcon  change the nav bar it is now resposnive to any size