landing page
Separate JavaScript Files for Country Utils:

Created a file countryUtils.js to contain utility functions related to country codes and names.
Added functions countryCodeToName and countryCodes to convert country codes to names and export them.
Integration of Country Utils:

Integrated the countryCodeToName function and countryCodes array into the main code to convert country codes to names.
Added Bahamas as a Country Code:

Added 'BHS' to the countryCodes array and included the corresponding name, "Bahamas", in the countryCodeToName function.
Removed Duplicate Countries:

Removed duplicate countries from the countryCodes array to ensure each country appears only once.
Implemented Search Field Zoom Behavior:

Modified the SearchField component to zoom in on the map only if the search result corresponds to a country.
Updated Map Zoom Behavior:

Adjusted the map zoom behavior so that it zooms in on the selected country when clicked, without marking it.